Refactored tool options to mix-ins #160
