+++ tests/cli/extraction_tool.py
@@ -6,7 +6,6 @@ import argparse
 import unittest
 
 from plaso.cli import extraction_tool
-from plaso.lib import errors
 
 from tests import test_lib as shared_test_lib
 from tests.cli import test_lib
@@ -154,14 +153,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
 
     test_tool._ParsePerformanceOptions(options)
 
-  def testParseStorageOptions(self):
-    """Tests the _ParseStorageOptions function."""
-    test_tool = extraction_tool.ExtractionTool()
-
-    options = test_lib.TestOptions()
-
-    test_tool._ParseStorageOptions(options)
-
   @shared_test_lib.skipUnlessHasTestFile([u'yara.rules'])
   def testParseYaraRulesOption(self):
     """Tests the _ParseYaraRulesOption function."""
@@ -273,21 +264,6 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
 
     expected_line = b'winreg_default : Parser for Registry data.'
     self.assertIn(expected_line, lines)
-
-  @shared_test_lib.skipUnlessHasTestFile([u'ímynd.dd'])
-  def testParseOptions(self):
-    """Tests the ParseOptions function."""
-    test_tool = extraction_tool.ExtractionTool()
-
-    options = test_lib.TestOptions()
-
-    # ParseOptions will raise if source is not set.
-    with self.assertRaises(errors.BadConfigOption):
-      test_tool.ParseOptions(options)
-
-    options.source = self._GetTestFilePath([u'ímynd.dd'])
-
-    test_tool.ParseOptions(options)
 
 
 if __name__ == '__main__':
